- Fixed casting warning - Formated main.cpp ..
git-svn-id: svn://svn.rocksteady-hosting.com/m2mp@20 8cfec0bd-d3a4-4003-b275-27d06a529d82

+++ trunk/Shared/SharedUtility.cpp
@@ -121,7 +121,7 @@ namespace SharedUtility
 
 	unsigned long GetTime( void )
 	{
-		return (std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::system_clock::now().time_since_epoch()).count());
+		return static_cast<unsigned int>((std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::system_clock::now().time_since_epoch()).count()));
 	}
 
 	bool Exists( const char * szPath )
